convert FormattedText in ts
