+++ package.json
@@ -40,6 +40,7 @@
   ],
   "homepage": "https://github.com/buildo/react-components",
   "dependencies": {
+    "@types/linkify-it": "2.0.1",
     "@types/lodash": "4.14.74",
     "@types/react": "15.6.1",
     "@types/react-dates": "^12.1.2",
