--- src/FormattedText/index.d.ts
@@ -1,2 +1,0 @@
-import FormattedText from './FormattedText';
-export default FormattedText;
