+++ src/FormattedText/index.ts
@@ -1,0 +1,3 @@
+import FormattedText, { FormattedTextProps } from './FormattedText';
+export default FormattedText;
+export { FormattedTextProps };
